+++ src/Particle.h
@@ -9,26 +9,26 @@ private:
 	static int pCount;
 public:
 	// Attributes of particle
-	float mass, size, elasticity;
-	glm::vec3 position, velocity, acceleration;
-	glm::vec3 force;
+	double mass, size, elasticity;
+	glm::dvec3 position, velocity, acceleration;
+	glm::dvec3 force;
 
 	// For linked list
 	Particle* next;
 
-	float density;
-	float pressure;
+	double density;
+	double pressure;
 	float id;
 	int type;//区分不同的粒子类型（这里只有两类）1,2
 	float ci;//color属性，ci=0.5,-0.5;
-	float gas_constant;
-	float viscosity;
-	float alpha;//DFSPH factor
+	double gas_constant;
+	double viscosity;
+	double alpha;//DFSPH factor
 	double densityAdv;
-	float kai;
+	double kai;
 	double rho_predict;
 
-	Particle(float mass, float size, glm::vec3 position, glm::vec3 velocity,int type,float ci);
+	Particle(double mass, float size, glm::dvec3 position, glm::dvec3 velocity,int type,float ci);
 	~Particle();
 };
